Rename networkParams to network-params
