+++ packages/claim-cli/src/applications/check-eligibility.ts
@@ -1,4 +1,4 @@
-import { NetworkParams } from '../utils/networkParams';
+import { NetworkParams } from '../utils/network-params';
 import buildAccountList, { AccountListChoice } from '../utils/build-account-list';
 import { fetchCheckEligibility } from '../utils/endpoint';
 import { getInput } from '../utils/get-prompts';
